Fikset actions under vis-alle, gjort start-korrigering gjenbrukbare og fikset rekkefølge på periodeliste visning
