+++ app/components/periodeliste-visning/PeriodelisteVisning.tsx
@@ -10,32 +10,6 @@ import {
 } from "~/models/rapporteringsperiode.server";
 import { type IRapporteringLoader } from "~/routes/rapportering";
 
-export async function action({ request }: ActionArgs) {
-  const formdata = await request.formData();
-  const submitKnapp = formdata.get("submit");
-
-  switch (submitKnapp) {
-    case "start-korrigering": {
-      const periodeId = formdata.get("periode-id") as string;
-
-      const korrigeringResponse = await startKorrigering(periodeId, request);
-
-      if (korrigeringResponse.ok) {
-        const korrigeringsperiode: IRapporteringsperiode = await korrigeringResponse.json();
-        return redirect(`/rapportering/endre/${korrigeringsperiode.id}`);
-      } else {
-        json({ korrigeringsfeil: true });
-      }
-    }
-
-    case "avgodkjenn": {
-      const periodeId = formdata.get("periode-id") as string;
-
-      return await avgodkjennPeriode(periodeId, request);
-    }
-  }
-}
-
 export default function PeriodelistVisning() {
   const { allePerioder } = useRouteLoaderData("routes/rapportering") as IRapporteringLoader;
 
@@ -50,7 +24,6 @@ export default function PeriodelistVisning() {
       {allePerioder?.map((periode: IRapporteringsperiode) => (
         <li key={periode.id}>
           {periode.fraOgMed} {periode.tilOgMed} - {periode.status} ({periode.id})
-          <Kalender rapporteringsperiode={periode} aapneModal={aapneModal} />
           <Form method="post">
             <input type="hidden" name="periode-id" value={periode.id}></input>
             {periode.status === "TilUtfylling" && (
@@ -69,6 +42,7 @@ export default function PeriodelistVisning() {
               </Button>
             )}
           </Form>
+          <Kalender rapporteringsperiode={periode} aapneModal={aapneModal} />
         </li>
       ))}
     </ul>
